requirements.txt
pandas==2.2.2
numpy==2.0.1
scikit-learn==1.5.1
kafka-python==2.0.2
streamlit==1.37.1
matplotlib==3.9.2
joblib==1.4.2
